Actor reference ID on debug window, RefID is now Int, added Japanese on help page
Former-commit-id: 3519b8746d611c4badd65f01644ba05e41d10d04
Former-commit-id: 45bb407c2d932e5d70aab9c9eb4f3ef55ce68d27

+++ src/net/torvald/terrarum/gameactors/Actor.kt
@@ -11,9 +11,9 @@ interface Actor {
 
     /**
      * Valid RefID is equal to or greater than 32768.
-     * @return Reference ID. (32768-0x7FFF_FFFF_FFFF_FFFF)
+     * @return Reference ID. (32768-0xFFFF_FFFF)
      */
-    var referenceID: Long
+    var referenceID: Int
 
     var actorValue: ActorValue
 }
